metrics: move event metrics into their own package
This commit moves event metrics into their own package.
