+++ pkg/metrics/eventmetrics/eventmetrics.go
@@ -1,0 +1,110 @@
+// SPDX-License-Identifier: Apache-2.0
+// Copyright Authors of Tetragon
+
+package eventmetrics
+
+import (
+	"strings"
+
+	v1 "github.com/cilium/hubble/pkg/api/v1"
+	"github.com/cilium/tetragon/api/v1/tetragon"
+	"github.com/cilium/tetragon/api/v1/tetragon/codegen/helpers"
+	"github.com/cilium/tetragon/pkg/api/processapi"
+	"github.com/cilium/tetragon/pkg/filters"
+	"github.com/cilium/tetragon/pkg/logger"
+	"github.com/cilium/tetragon/pkg/metrics/consts"
+	readerdns "github.com/cilium/tetragon/pkg/reader/dns"
+	"github.com/cilium/tetragon/pkg/reader/exec"
+	"github.com/prometheus/client_golang/prometheus"
+	"github.com/prometheus/client_golang/prometheus/promauto"
+)
+
+var (
+	eventsProcessed = promauto.NewCounterVec(prometheus.CounterOpts{
+		Name:        consts.MetricNamePrefix + "events_total",
+		Help:        "The total number of Tetragon events",
+		ConstLabels: nil,
+	}, []string{"type", "namespace", "pod", "binary"})
+	flagCount = promauto.NewCounterVec(prometheus.CounterOpts{
+		Name:        consts.MetricNamePrefix + "flags_total",
+		Help:        "The total number of Tetragon flags. For internal use only.",
+		ConstLabels: nil,
+	}, []string{"type"})
+	dnsRequestTotal = promauto.NewCounterVec(prometheus.CounterOpts{
+		Name: consts.MetricNamePrefix + "dns_total",
+		Help: "Dns request/response statistics",
+	}, []string{"namespace", "pod", "binary", "names", "rcodes", "response"})
+)
+
+func getProcessInfo(process *tetragon.Process) (binary, pod, namespace string) {
+	if process != nil {
+		binary = process.Binary
+		if process.Pod != nil {
+			namespace = process.Pod.Namespace
+			pod = process.Pod.Name
+		}
+	}
+	return binary, pod, namespace
+}
+
+func handleOriginalEvent(originalEvent interface{}) {
+	var flags uint32
+	switch msg := originalEvent.(type) {
+	case *processapi.MsgExecveEventUnix:
+		flags = msg.Process.Flags
+	}
+	for _, flag := range exec.DecodeCommonFlags(flags) {
+		flagCount.WithLabelValues(flag).Inc()
+	}
+}
+
+func postDnsMetric(ev *tetragon.GetEventsResponse, res *tetragon.ProcessDns) {
+	var rr string
+
+	binary, pod, ns := getProcessInfo(filters.GetProcess(&v1.Event{Event: ev}))
+
+	dns := res.Dns
+	names := strings.Join(dns.GetNames(), ",")
+	codes := readerdns.GetRCodeString(uint16(dns.GetRcode()))
+
+	if dns.Response {
+		rr = "Response"
+	} else {
+		rr = "Request"
+	}
+
+	dnsRequestTotal.WithLabelValues(ns, pod, binary, names, codes, rr).Inc()
+}
+
+func handleDnsEvent(processedEvent interface{}) {
+	switch ev := processedEvent.(type) {
+	case *tetragon.GetEventsResponse:
+		switch res := ev.Event.(type) {
+		case *tetragon.GetEventsResponse_ProcessDns:
+			postDnsMetric(ev, res.ProcessDns)
+		}
+	}
+}
+
+func handleProcessedEvent(processedEvent interface{}) {
+	var eventType, namespace, pod, binary string
+	switch ev := processedEvent.(type) {
+	case *tetragon.GetEventsResponse:
+		binary, pod, namespace = getProcessInfo(filters.GetProcess(&v1.Event{Event: ev}))
+		var err error
+		eventType, err = helpers.EventTypeString(ev.Event)
+		if err != nil {
+			logger.GetLogger().WithField("event", processedEvent).WithError(err).Warn("metrics: handleProcessedEvent: unhandled event")
+			eventType = "unhandled"
+		}
+	default:
+		eventType = "unknown"
+	}
+	eventsProcessed.WithLabelValues(eventType, namespace, pod, binary).Inc()
+}
+
+func ProcessEvent(originalEvent interface{}, processedEvent interface{}) {
+	handleOriginalEvent(originalEvent)
+	handleProcessedEvent(processedEvent)
+	handleDnsEvent(processedEvent)
+}
